Fix the string format issue
Summary:

mac and our dev server has totally differnt definition of uint64_t, therefore fixing the warning in mac has actually made code in linux uncompileable.

Test Plan:

make clean && make -j32

+++ db/memtablelist.cc
@@ -122,7 +122,9 @@ Status MemTableList::InstallMemtableFlushResults(
       break;
     }
 
-    Log(info_log, "Level-0 commit table #%llu started", m->file_number_);
+    Log(info_log,
+        "Level-0 commit table #%lu started",
+        (unsigned long)m->file_number_);
 
     // this can release and reacquire the mutex.
     s = vset->LogAndApply(&m->edit_, mu);
@@ -133,9 +135,9 @@ Status MemTableList::InstallMemtableFlushResults(
     do {
       if (s.ok()) { // commit new state
         Log(info_log,
-            "Level-0 commit table #%llu: memtable #%llu done",
-            m->file_number_,
-            mem_id);
+            "Level-0 commit table #%lu: memtable #%lu done",
+            (unsigned long)m->file_number_,
+            (unsigned long)mem_id);
         memlist_.remove(m);
         assert(m->file_number_ > 0);
 
@@ -149,9 +151,9 @@ Status MemTableList::InstallMemtableFlushResults(
       } else {
         //commit failed. setup state so that we can flush again.
         Log(info_log,
-            "Level-0 commit table #%llu: memtable #%llu failed",
-            m->file_number_,
-            mem_id);
+            "Level-0 commit table #%lu: memtable #%lu failed",
+            (unsigned long)m->file_number_,
+            (unsigned long)mem_id);
         m->flush_completed_ = false;
         m->flush_in_progress_ = false;
         m->edit_.Clear();
